Activities v2

+++ apps/api/src/modules/activities/activities.controller.ts
@@ -7,7 +7,8 @@ import {
   Param,
   Delete,
   Query,
-  UseGuards
+  UseGuards,
+  ForbiddenException
 } from '@nestjs/common'
 import { ActivitiesIminService } from './activities.imin.service'
 import { ActivitiesService } from './activities.service'
@@ -17,12 +18,15 @@ import {
   FindActivitiesDto,
   FindActivitiesForMapDto
 } from './dto/find-activities.dto'
-import { Pagination } from '../../helpers/paginate'
+import { Pagination, PaginationQuery } from '../../helpers/paginate'
 import { Activity, ActivityForMap } from './entities/activity.entity'
 import { Public } from '../../decorators/public.decorator'
 import { ApiTags } from '@nestjs/swagger'
 import { IminServiceParams } from './types/imin'
 import { ApiBaseResponses } from '../../decorators/swagger.decorator'
+import { User } from '../../decorators/authenticated-user.decorator'
+import { Pagination as PaginationParams } from '../../decorators/pagination.decorator'
+import { AuthenticatedUser } from '../../models'
 
 // @Public()
 @ApiBaseResponses()
@@ -33,10 +37,12 @@ export class ActivitiesController {
     private readonly activitiesService: ActivitiesService,
     private readonly activitiesIminService: ActivitiesIminService
   ) {}
-
   @Post('/activities')
-  async create(@Body() dto: CreateActivityDto) {
-    return this.activitiesService.create(dto)
+  async create(
+    @User() user: AuthenticatedUser,
+    @Body() dto: CreateActivityDto
+  ) {
+    return this.activitiesService.create(user.id, dto)
   }
 
   /**
@@ -45,12 +51,13 @@ export class ActivitiesController {
    * @param queryParams
    * @returns
    */
-  @Get('/activities/user/:userId')
-  async findUserActivities(@Param('userId') userId: string, @Query() query) {
-    return this.activitiesService.findUserActivities(userId, {
-      limit: parseInt(query.limit) || 10,
-      page: parseInt(query.page) || 0
-    })
+  @ApiTags('me')
+  @Get('/me/activities')
+  async findUserActivities(
+    @User() user: AuthenticatedUser,
+    @PaginationParams() pagination: PaginationQuery
+  ) {
+    return this.activitiesService.findUserActivities(user.id, pagination)
   }
 
   /**
@@ -62,6 +69,7 @@ export class ActivitiesController {
    */
   @Get('/activities/map')
   async findAllForMap(
+    @User() user: AuthenticatedUser,
     @Query()
     {
       geo_radial,
@@ -72,6 +80,7 @@ export class ActivitiesController {
   ): Promise<Pagination<ActivityForMap>> {
     // Get local activities
     const all = await this.activitiesService.findAllMarkers(
+      user.id,
       geo_radial,
       type,
       keyword
@@ -117,6 +126,7 @@ export class ActivitiesController {
    */
   @Get('/activities')
   async findAll(
+    @User() user: AuthenticatedUser,
     @Query()
     {
       geo_radial,
@@ -132,6 +142,7 @@ export class ActivitiesController {
 
     // Get local activities
     const all = await this.activitiesService.findAll(
+      user.id,
       geo_radial,
       type,
       keyword,
@@ -170,43 +181,51 @@ export class ActivitiesController {
   }
 
   @Get('/activities/:id')
-  findOne(@Param('id') id: string) {
+  findOne(@User() user: AuthenticatedUser, @Param('id') id: string) {
     if (id.length === 36) {
-      return this.activitiesService.findOne(id)
+      if (user.isSuperAdmin()) {
+        return this.activitiesService.findOne(id)
+      }
+      return this.activitiesService.findOne(id, user.id)
     } else {
       return this.activitiesIminService.findOne(id)
     }
   }
 
-  @Get('/activities/:userId/:id')
-  findOneUserActivity(
-    @Param('id') id: string,
-    @Param('userId') userId: string
-  ) {
-    return this.activitiesService.findOneUserActivity(id, userId)
-  }
-
   @Put('/activities/:id')
   async update(
+    @User() user: AuthenticatedUser,
     @Param('id') id: string,
     @Body() updateActivityDto: UpdateActivityDto
   ) {
-    return this.activitiesService.update(id, updateActivityDto)
+    if (user.isSuperAdmin()) {
+      return this.activitiesService.update(id, updateActivityDto)
+    } else {
+      const activity = this.activitiesService.findOneUserActivity(id, user.id)
+      if (activity) {
+        return this.activitiesService.update(id, updateActivityDto)
+      } else {
+        throw new ForbiddenException(
+          'You do not have access to update this activity'
+        )
+      }
+    }
   }
 
   @Delete('/activities/:id')
-  remove(@Param('id') id: string) {
-    return this.activitiesService.remove(id)
-  }
-
-  @Delete('/activities/:id/images')
-  removeImages(@Param('id') id: string) {
-    return this.activitiesService.removeImages(id, 'images')
-  }
-
-  @Delete('/activities/:id/organizer_image')
-  removeOrganizerImage(@Param('id') id: string) {
-    return this.activitiesService.removeImages(id, 'organizer_image')
+  remove(@User() user: AuthenticatedUser, @Param('id') id: string) {
+    if (user.isSuperAdmin()) {
+      return this.activitiesService.remove(id)
+    } else {
+      const activity = this.activitiesService.findOneUserActivity(id, user.id)
+      if (activity) {
+        return this.activitiesService.remove(id)
+      } else {
+        throw new ForbiddenException(
+          'You do not have access to update this activity'
+        )
+      }
+    }
   }
 
   static mergeAndPaginate(
